More explicit naming of 16 bit integers.
git-svn-id: svn://scm.orgis.org/mpg123/trunk@3512 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/sample.h
@@ -14,11 +14,11 @@
 
 /* Special case is fixed point math... which does work, but not that nice yet.  */
 #ifdef REAL_IS_FIXED
-static inline short idiv_signed_rounded(long x, int shift)
+static inline int16_t idiv_signed_rounded(int32_t x, int shift)
 {
 	x >>= (shift - 1);
 	x += (x & 1);
-	return (short)(x >> 1);
+	return (int16_t)(x >> 1);
 }
 #  define REAL_PLUS_32767       ( 32767 << 15 )
 #  define REAL_MINUS_32768      ( -32768 << 15 )
@@ -36,7 +36,7 @@ static inline short idiv_signed_rounded(long x, int shift)
 # if (defined REAL_IS_FLOAT) && (defined IEEE_FLOAT)
 /* This function is only available for IEEE754 single-precision values
    This is nearly identical to proper rounding, just -+0.5 is rounded to 0 */
-static inline short ftoi16(float x)
+static inline int16_t ftoi16(float x)
 {
 	union
 	{
@@ -44,7 +44,7 @@ static inline short ftoi16(float x)
 		int32_t i;
 	} u_fi;
 	u_fi.f = x + 12582912.0f; /* Magic Number: 2^23 + 2^22 */
-	return (short)u_fi.i;
+	return (int16_t)u_fi.i;
 }
 #  define REAL_TO_SHORT_ACCURATE(x)      ftoi16(x)
 # else
@@ -139,7 +139,7 @@ static inline short ftoi16(float x)
 /* Produce an 8bit sample, via 16bit intermediate. */
 #define WRITE_8BIT_SAMPLE(samples,sum,clip) \
 { \
-	short write_8bit_tmp; \
+	int16_t write_8bit_tmp; \
 	if( (sum) > REAL_PLUS_32767) { write_8bit_tmp = 0x7fff; (clip)++; } \
 	else if( (sum) < REAL_MINUS_32768) { write_8bit_tmp = -0x8000; (clip)++; } \
 	else { write_8bit_tmp = REAL_TO_SHORT(sum); } \
